pyproject.toml
[project]
name = "adversarialattack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.116.1",
    "matplotlib>=3.9.4",
    "pillow>=11.3.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "streamlit>=1.48.1",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "uvicorn>=0.35.0",
]
